requirements.txt
bcrypt
certifi
cffi
chardet
click
cryptography
Flask
idna
itsdangerous
Jinja2
junos-eznc
lxml
MarkupSafe
ncclient
netaddr
paramiko
python-gitlab
PyYAML
requests
scp
six
urllib3
wakeonlan
Werkzeug
uwsgi